Extensions: Move extensions-middleware to separate directory
